devicemonitoring/requirements.txt
fastapi
pydantic
pymongo
paho-mqtt
celery
redis
uvicorn
python-dotenv
gmqtt
